requirements.txt
Flask
Flask-Cors
python-dotenv
langchain-community
openai
Gunicorn